requirements.txt
absl-py==0.13.0
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.3.1
astroid==2.5.1
attrs==21.2.0
bloom-filter==1.3.3
BloomFilter==0.1.0
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
isort==5.8.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
mediapipe==0.8.7.2
mmh3==3.0.0
mysqlclient==2.0.3
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
packaging==21.2
Pillow==8.3.2
protobuf==3.17.3
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.3.0
pylint==2.7.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.5.66
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.4.3
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
